requirements.txt
Django==4.2.10
celery==5.3.6
django-celery-results==2.5.1
djangorestframework==3.14.0
django-filter==23.5
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
feedparser==6.0.11
pytest==8.0.1
pytest-django==4.8.0
pytest-mock==3.12.0
redis==5.0.1